Remove old dependency code
